+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 8b83d53 - v13
+| Author: fjrodafo - Wed, 11 Dec 2024 16:54:54 +0000
+| 
 * Commit: 3b7a834 - v12
 | Author: fjrodafo - Sat, 7 Dec 2024 19:08:38 +0000
 |